Test (#9)
* Remove Test page and update Index and layout templates

The Test page was deleted, and its content, including SEO metadata and descriptions, was moved to the Index page. Layout adjustments were made, including the removal of the Test page link, an updated copyright year, and modification of logging levels to Debug in the development environment.

* Refactor to add multilingual support with 'en' and 'nl' controllers.

Replaced 'HomeController' with 'enController' and added a new 'nlController' to support multilingual pages. Updated routing, views, and layout to reflect these changes and ensure proper navigation for both languages.

---------

Co-authored-by: nvoutersterp <nvoutersterp@gmail.com>